chore: update version to 0.0.6 and adjust SDK constraints
This commit updates the version to 0.0.6 and modifies the SDK constraints to support a wider range of Flutter versions. Additionally, it enhances the example app by renaming the main page and adding a comprehensive list of example components, improving the overall user experience. The Info.plist file is also updated to reflect the new app name and includes local network usage descriptions for debugging features.